Perks now exist
Perks are essentially a condition, they don't do anything on their own.

Perk.CanUse returns true if the perk is not on cooldown, and then starts the cooldown (if it has one).

Player has a Perk Controller that contains the player's active perks.
Use perkController.CanUse("perk id"). to check if the perk is active and isn't on cooldown.

Example use:

Weapon Fire 1()
{
    if (perkController.CanUse("doubleTap"))
        {
            // Shoot two bullets.
        }
}