Colour shifting
When the player takes damage, chromatic aberration is set to full. (Feedback?) - This will also trigger when the player gains health as for now, it triggers when the player's health changes.

Laid out the scripts that allow Text to be colour shifted. I just need to implement the way to trigger them when the player loses health.