Added reloading
Added reloading to weapons (can choose the reload delay, and how many bullets to fill (-1 = the max size).

Can't cancel the reload (is this important?)

Added an auto-reload option.

Weapons are automatically filled to max at start.

Added a nicer way to tell if the weapon is being controlled by the player. Just use the IsPlayer condition.