Weapon changes, AI delay, new perk.
Decreased the fire rate of all the weapons (seemed too fast for a bullet hell).

Added a delay when the player leaves an enemy so it doesn't instantly start shooting you (in the CharacterMotor class).

Added colour to each of the weapons so you can tell which weapon is what.

Added a light to my scene so the weapon colours showed properly.

Added the Juggernaught perk (whenever the player takes over a NEW enemy, the health is set to max).